requirements-dev.txt
# python3.10 | pythonanywhere
flask>=2.3.0,<3.0.0
flask-assets==2.0
flask-login>=0.6.0
flask-sqlalchemy>=3.0.0,<4.0.0
flask-wtf>=1.0.0
pillow>=9.0.0,<10.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
gitpython==3.1.43
email_validator==2.2.0
dnspython==2.7.0
sqlalchemy==2.0.36
werkzeug==3.1.3
wtforms==3.2.1
